refactor rebalancing, addressing bugs & improving benchmarking
Banged my head against a number of things here, and this still isn't
exactly what I want but I think it's usable or at least moves the ball
forward a bit.

I've been comparing the implementation here to django-lexorank &
lexorank-ts, which is helpful in some ways but all the implementations
all seem to conceptualize lexorank a bit differently so their
increment/rebalance approaches aren't trivially translateable either.

I think to really do this fully correctly the implementation here would
probably have to grow to consider buckets and things, which I've been
kind of trying to avoid.

A few stumbling blocks I hit:

- carry over in addition/incrementing. The basic idea here seems to be
  treating each rank as a base-n number. The django-lexorank
  implementation includes a carry-over at the end to prepend new
  "digits", but I think this is buggy since lexorank sorting doesn't
  work like normal numerical sorting. E.g. if you increment a key like
  `gT` a large amount and it carries over, it might end up as like
  `17R`, which is more "digits" but sorts earlier by string semantics.

  What I ended up with is a check to treat any leftover carry during
  increment as a bug. I think the approach to choosing the length of
  keys and the increment between them ensures there shouldn't in
  practice be any leftover during incrementing.
- end of keyspace being inclusive or exclusive. In other implementations
  I've been reading, 'z' or 'zzzz' or whatever would be a valid key, but
  this implementation is treating 'z' as an unreachable "infinity" in
  the keyspace, so I adjusted the step calculation to adjust for that.
  FWIW i think in practice the implementation here would handle keys
  above 'z' fine, so this might be an artificial limitation that's not
  really needed? I don't want to get sidetracked by rewriting
  fundamental behavior here, though.
- batches. By default batches in Rails ignore any existing `order` on
  your query and use primary key, I'd forgotten that, that created some
  bad bugs with things ending up in the wrong order after rebalancing.
  I'm also not 100% convinced this algorithm couldn't "leapfrog"
  existing records in a later batch during the key incrementing, so I
  moved from updating each batch as it's fetched to first calculating
  the entire set of new ranks in memory and then doing the upserts in
  batches. I'm dissatisfied with this since it has some performance
  risks, but in my benchmarking (see discussion below) I think it's
  tolerable for realistic use. I think the "proper" way of doing it
  means adding buckets, which, again, I'm trying not to do.

\## Benchmarking

The benchmark script here is both a perf test tool & a bit of a fuzz
tester on correctness since a unit test rebalancing large tables and
ensuring order is correct is probably not something you'd want running
every time you run the test suite.

Perf here feels pretty acceptable. Time to rebalance a 10k item table is
~300-400ms on my laptop. A 100k item table is ~3s. Realistic usage will
probably be a bit worse since this is an in-memory sqlite database and
latency with postgres/mysql is likely to be the bottleneck in real
usage, but for reasonably sized tables I think it's fast enough that the
duration of the operation (and the lock) is unlikely to be a big issue.

Memory usage is of course a bit higher since I'm building the full set
of updates in memory, but the fact that we only select id and rank helps
there. Even with 100k items, memory usage of the benchmark script was
only 50-something MB, so I think it's acceptable.